# [Enhancement] support per bucket optimize for colocate aggregate