… be/src/exec/CMakeLists.txt
… be/src/exec/pipeline/pipeline.cpp
… be/src/exec/pipeline/pipeline_driver.cpp
… be/src/exec/pipeline/scan/chunk_source.cpp
… be/src/exec/pipeline/scan/morsel.cpp
… be/src/exec/pipeline/scan/morsel.h
… be/src/exec/pipeline/scan/scan_operator.cpp
… be/src/exec/pipeline/scan/scan_operator.h
… be/src/exec/query_cache/cache_operator.h
… be/src/exec/query_cache/multilane_operator.h
… be/src/exec/query_cache/ticket_checker.cpp
… be/src/exec/query_cache/ticket_checker.h
… be/src/exec/scan_node.cpp
… be/src/exec/scan_node.h
… be/src/exec/spill/input_stream.h
… be/src/exec/spill/spiller.cpp
… be/src/exec/spill/spiller.h
… fe/fe-core/src/main/java/com/starrocks/planner/AggregationNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/OlapScanNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/PlanFragment.java
… fe/fe-core/src/main/java/com/starrocks/planner/PlanNode.java
… fe/fe-core/src/main/java/com/starrocks/qe/CoordinatorPreprocessor.java
… fe/fe-core/src/main/java/com/starrocks/qe/SessionVariable.java
… fe/fe-core/src/main/java/com/starrocks/qe/scheduler/assignment/LocalFragmentAssignmentStrategy.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Optimizer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalHashAggregateOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalOlapScanOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/AddDecodeNodeForDictStringRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/PhysicalDistributionAggOptRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/tree/UseSortAggregateRule.java