DOP | baseline | Peak mem | bucket optimize | Peak mem | ratio:
-- | -- | -- | -- | -- | --
16 | 2.42 sec | 9.308 GB | 1.93 sec | 2.368 GB |  
8 | 4.33 sec | 9.278 GB | 3.53 sec | 1.181 GB |  
4 | 8.25 sec | 9.266 GB | 6.28 sec | 606.039 MB |  
1 | 33.02 sec | 9.257 GB | 23.86 sec | 154.247 MB |  

compare to sorted streaming agg.
DOP | Sort AGG | Bucket compute | ratio:
-- | -- | -- | --
1 | 9.12 sec | 24.18 sec |  

It looks like there is some room for optimization. I will do it in the future.


## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4