@wanpengfei-git 2023-08-15
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 82 / 89 (92.13%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/OlapScanNode.java|8|10|80.00%|[407, 408]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/tree/PhysicalDistributionAggOptRule.java|46|51|90.20%|[45, 92, 97, 102, 106]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalOlapScanOperator.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/AggregationNode.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalHashAggregateOperator.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/tree/AddDecodeNodeForDictStringRule.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/CoordinatorPreprocessor.java|6|6|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/ColocatedBackendSelector.java|2|2|100.00%|[]|