@fzhedu 2023-08-17
> if there are too many small tables, should we merge several one to comput?

ignore it