@wanpengfei-git 2023-08-18
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 89 / 95 (93.68%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/OlapScanNode.java|9|11|81.82%|[404, 405]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/tree/PhysicalDistributionAggOptRule.java|47|51|92.16%|[45, 92, 97, 102]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalOlapScanOperator.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/AggregationNode.java|7|7|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalHashAggregateOperator.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/tree/AddDecodeNodeForDictStringRule.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/PlanFragment.java|7|7|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/scheduler/assignment/LocalFragmentAssignmentStrategy.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|1|1|100.00%|[]|